♻️ 채널 쿼리 훅 리팩토링 (#269)
